Update "Normal" bars.ipynb
Time sampling creates a lot of rows containing NaNs as the sampling occurs over the whole day and not just during the trading hours. When plotting the histogram of the bar-on-bar price changes these NaNs are automatically converted to ZERO thus disproportionately increasing the number of observations falling the the zero bin. By removing the observations containing the NaNs the histogram will "behave" better than before.